refactor: relocate ClientProviders to a dedicated providers directory for better organization
